style: Modify the framework of Assets file reading and add the logic layer of Assets.
Modify the framework of Assets file reading and add the logic layer of
Assets.
