+++ gui/libDataConvert/CMakeLists.txt
@@ -12,6 +12,7 @@ add_library(libDataConvert
     include/data_Convert/ProgramManager.h src/ProgramManager.cpp
     include/data_Convert/OutputAssets.h src/OutputAssets.cpp
     include/data_Convert/InputAssets.h src/InputAssets.cpp
+    include/data_Convert/AssetsLogic.h src/AssetsLogic.cpp
     include/proto/Common.pb.h src/Common.pb.cpp
     include/proto/HmiBase.pb.h src/HmiBase.pb.cpp
     include/proto/HmiWidget.pb.h src/HmiWidget.pb.cpp
