+++ gui/libDataConvert/include/data_Convert/AssetsLogic.h
@@ -1,5 +1,5 @@
-#ifndef INPUT_ASSETS_H
-#define INPUT_ASSETS_H
+#ifndef ASSETS_LOGIC_H
+#define ASSETS_LOGIC_H
 
 #include "PropertyData/PropertyData.h"
 #include "NodeData/nodeManager.h"
@@ -7,7 +7,9 @@
 #include "AnimationData/animationData.h"
 #include "CurveData/CurveManager.h"
 #include "MaterialData/materialManager.h"
+#include "signal/SignalProxy.h"
 #include "MeshData/MeshDataManager.h"
+#include "data_Convert/InputAssets.h"
 
 #include "proto/Numeric.pb.h"
 #include "proto/Common.pb.h"
@@ -19,20 +21,24 @@
 #include <stdio.h>
 #include <string>
 #include <iostream>
-
+#include <fstream>
+#include <qdir.h>
 
 using namespace raco::guiData;
 
 namespace raco::dataConvert {
-class InputPtx {
+class AssetsLogic : public QObject {
+    Q_OBJECT
 public:
-	bool parseOneNode(const HmiScenegraph::TNode& tNode, NodeData& nodeData);
+	bool readProgram2Ptx(std::string filePathStr);
 	
+Q_SIGNALS:
 
 private:
-
+	bool isInputError_{false};
+	InputPtx inputPtx_;
 };
 
 }
 
-#endif // INPUT_ASSETS_H
+#endif // ASSETS_LOGIC_H
